Folder Structure : - restructure Folder tree, to make the project easier to maintain
