+++ lib/src/common/constants/empty_widget.dart
@@ -1,6 +1,6 @@
 import 'package:books_app/src/common/constants/asset_path.dart';
 import 'package:books_app/src/common/constants/colors.dart';
-import 'package:books_app/src/common/widgets/typography.dart';
+import 'package:books_app/src/presentation/widgets/typography.dart';
 import 'package:flutter/material.dart';
 
 class EmptyWidget extends StatelessWidget {
